Script to fix up the Library Reference index so it's less ugly.  The docstring:
"""Combine similar index entries into an entry and subentries.

For example:

  \item {foobar} (in module flotz), 23
  \item {foobar} (in module whackit), 4323

becomes

  \item {foobar}
    \subitem in module flotz, 23
    \subitem in module whackit, 4323

Note that an item which matches the format of a collapsable item but which
isn't part of a group of similar items is not modified.
"""

This results in a much more readable index, with less repitition of text;
especially for common method names.

--HG--
branch : legacy-trunk