Lots of nitty-gritty stuff:
Use \emph{} when referring to title of any of the Python manuals, like is
done in the other manuals.

In some places, use \emph{} or \dfn{} instead of ``...''.

Use \samp{} instead of \code{} when marking identifier prefixes.

Use logical markup wherever it made sense.

Fixed a bunch of typos.

In several places, use "---" instead of "--" to get the emdash.

Start sentences with capital letters and end them with periods, as needed.

"it's" --> "its" in many places:  "it's" it *always* a contraction of "it is",
and "its" should always be used for the possessive.

"don't" --> "doesn't"

"should discards any" --> "should discard any"

In C function descriptions, use \var{} to mark parameters in the running
text instead of \code{}.  This matches usage in the other manauls better,
and is more consistent with the formatting of the function signatures.
Lots of little changes to implement this.

Lots of fixups needed due to appearant heavy cut-&-paste in the orignal
document.

Mention that the exception objects may be either classes or strings,
depending on the use of -X; they were unequivocally stated to be strings
in the section "Standard Exceptions".

"mkvalue()" --> "Py_BuildValue()"

Description of PyNumber_Power() indicates that the third value is option,
but not how to indicate that it was omitted.  Clarified.

Explain the behavior of PyString_FromStringAndSize() if the buffer is NULL.

Explain the Py_complex structure a little (tell what it's used for) and use
a {verbatim} environment for the structure definition itself.

Fix explanation of PyFile_SoftSpace().

Update the example version string to 1.5.

Combined the sections on defining new object types.

--HG--
branch : legacy-trunk