Change the date to March 6, 1998.
Migration to Frame 5.5.

Checking in the PDF file.  (This is questionable -- but we check in the
Postscript as well.  Perhaps *both* should be removed from the CVS tree,
now everything is unbundled and we're distributing PS and PDF tarballs?)

Note -- the HTML generation using webmaker is currently broken.  I'll
have to look into generating HTML from Frame...

--HG--
branch : legacy-trunk