Don't use \samp when \code is better.
Consistently use ' instead of ":  this conforms to my unofficial stance that
constants should be presented using the repr() whenever it makes sense.  This
isn't because I think repr() is the greatest thing since spam, just that it's
an easy to adopt consistency standard.  (It also holds no new surprises!)

--HG--
branch : legacy-trunk