(python-mode): Set which interpreter (CPython or JPython) to use the first time a py buffer is visited during the Emacs session.  This ensures that py-which-shells is initialized and also guarantees that the mode lines reflect the correct shell.  First bug found by GvR, second one has long bugged :) me.
(py-toggle-shells): Programmatically, arg can also take the symbols
`cpython' or `jpython', which makes it easy to call with the value of
py-default-interpreter.

(py-shell): Don't need to initialize py-which-* variables since these
will guarantee to be initialized by python-mode when the first py
buffer is visited.

(py-default-interpreter): Update docstring.

--HG--
branch : legacy-trunk